# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Waste Platform API",
    "description": "منصة إدارة النفايات الصلبة وإعادة التدوير — REST API",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/requests": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "lines"
                ],
                "properties": {
                  "notes": {
                    "maxLength": 500,
                    "type": "string"
                  },
                  "lines": {
                    "minItems": 1,
                    "maxItems": 50,
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "wasteTypeCode"
                      ],
                      "properties": {
                        "wasteTypeCode": {
                          "minLength": 2,
                          "maxLength": 32,
                          "type": "string"
                        },
                        "quantity": {
                          "pattern": "^\\d+(\\.\\d{1,3})?$",
                          "type": "string"
                        },
                        "weightKg": {
                          "pattern": "^\\d+(\\.\\d{1,3})?$",
                          "type": "string"
                        },
                        "selectedAddons": {
                          "maxItems": 10,
                          "type": "array",
                          "items": {
                            "maxLength": 32,
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "get": {
        "parameters": [
          {
            "schema": {
              "minimum": 1,
              "maximum": 10000,
              "type": "integer"
            },
            "in": "query",
            "name": "page",
            "required": false
          },
          {
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "type": "integer"
            },
            "in": "query",
            "name": "pageSize",
            "required": false
          },
          {
            "schema": {
              "maxLength": 32,
              "type": "string"
            },
            "in": "query",
            "name": "status",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/requests/{id}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/requests/{id}/transitions": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "target",
                  "expectedVersion"
                ],
                "properties": {
                  "target": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "received"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "sent_to_collector"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "on_the_way"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "arrived"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "collected"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "sorted"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "sold"
                        ]
                      }
                    ]
                  },
                  "expectedVersion": {
                    "minimum": 1,
                    "type": "integer"
                  },
                  "barcode": {
                    "maxLength": 200,
                    "type": "string"
                  },
                  "collectorUserId": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "scheduledDay": {
                    "maxLength": 20,
                    "type": "string"
                  },
                  "scheduledHour": {
                    "maxLength": 10,
                    "type": "string"
                  },
                  "reason": {
                    "minLength": 3,
                    "maxLength": 300,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
}